Moved application functions to profile store and deleted application store
